example version rollback

+++ examples/get_started/lv_example_get_started_3.c
@@ -8,8 +8,8 @@ static void slider_event_cb(lv_event_t * e)
     lv_obj_t * slider = lv_event_get_target(e);
 
     /*Refresh the text*/
-    lv_label_set_text_fmt(label, "%d", (int)lv_slider_get_value(slider));
-    lv_obj_align_to(label, slider, LV_ALIGN_OUT_TOP_MID, 0, -15);    /*Align top of the slider*/
+    lv_label_set_text_fmt(label, "%d", lv_slider_get_value(slider));
+    lv_obj_align_to(label, slider, LV_ALIGN_OUT_TOP_MID, 0, -15);    /*Align below the slider*/
 }
 
 /**
@@ -26,7 +26,7 @@ void lv_example_get_started_3(void)
     /*Create a label below the slider*/
     label = lv_label_create(lv_scr_act());
     lv_label_set_text(label, "0");
-    lv_obj_align_to(label, slider, LV_ALIGN_OUT_TOP_MID, 0, -15);    /*Align top of the slider*/
+    lv_obj_align_to(label, slider, LV_ALIGN_OUT_TOP_MID, 0, -15);    /*Align below the slider*/
 }
 
 #endif
